+++ src/importer/components-map.js
@@ -36,8 +36,8 @@ export function build(targetComponentsDir: string): Promise<Object> {
     glob('*/*/*/*', { cwd: targetComponentsDir }, (err, files) => {
       if (err) return reject(err);
       files.forEach((loc) => {
-        const [box, name, scope, version] = loc.split(path.sep);
-        const id = scope + ID_DELIMITER + box + ID_DELIMITER + name + VERSION_DELIMITER + version;
+        const [namespace, name, scope, version] = loc.split(path.sep);
+        const id = scope + ID_DELIMITER + namespace + ID_DELIMITER + name + VERSION_DELIMITER + version;
         const bitJson = BitJson.load(path.join(targetComponentsDir, loc));
         const dependencies = [];
         Object.keys(bitJson.dependencies).forEach((dependency) => {